agentic-os/backend/pyproject.toml
[project]
name = "solvin"
version = "0.1"
description = "Solvin Agentic Operating System CLI - Unified entrypoint for all microservices"
readme = "README.md"
authors = [
  { name = "Yaniv Shemesh", email = "y@solvin.ai" }
]
requires-python = ">=3.12"
dependencies = [
  "typer>=0.15.3",
  "click>=8.2.0",
  "requests>=2.32.3",
  "python-json-logger>=3.3.0",
  "pyyaml>=6.0.2"
]

[project.scripts]
solvin = "solvin.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
  "solvin",
  "modules",     
  "modules.*", 
  "shared"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
